@@ -138,8 +138,7 @@ final class EntityNormalizer extends ObjectNormalizer
      */
     private function handleFotoPathes($object, array $data): array
     {
-        if (
-            Koopman::class === get_class($object) ||
+        if (Koopman::class === get_class($object) ||
             Vervanger::class === get_class($object)
         ) {
             $fotoUrl = null;
@@ -165,8 +164,7 @@ final class EntityNormalizer extends ObjectNormalizer
      */
     private function handleShyProperties($object, array $data): array
     {
-        if (
-            Dagvergunning::class === get_class($object) &&
+        if (Dagvergunning::class === get_class($object) &&
             array_key_exists('controles', $data) &&
             is_array($data['controles']) &&
             count($data['controles']) < 1
@@ -193,8 +191,7 @@ final class EntityNormalizer extends ObjectNormalizer
      */
     private function handleMarktIndelingslijst($object, array $data): array
     {
-        if (
-            Markt::class === get_class($object)
+        if (Markt::class === get_class($object)
         ) {
             $data['isABlijstIndeling'] = ('A/B-lijst' === $object->getIndelingstype());
         }
